pyproject.toml
[project]
name = "filesystem-server"
version = "0.1.0"
description = "MCP server for filesystem access with configurable directory and file type restrictions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
]